+++ shiny/templates/chat/hello-providers/anthropic/app.py
@@ -1,14 +1,23 @@
 # ------------------------------------------------------------------------------------
 # A basic Shiny Chat example powered by Anthropic's Claude model.
-# To run it, you'll need an Anthropic API key.
-# To get one, follow the instructions at https://docs.anthropic.com/en/api/getting-started
 # ------------------------------------------------------------------------------------
 import os
 
 from app_utils import load_dotenv
-from chatlas import AnthropicChat
+from chatlas import ChatAnthropic
 
 from shiny.express import ui
+
+# ChatAnthropic() requires an API key from Anthropic.
+# See the docs for more information on how to obtain one.
+# https://posit-dev.github.io/chatlas/reference/ChatAnthropic.html
+load_dotenv()
+chat_model = ChatAnthropic(
+    api_key=os.environ.get("ANTHROPIC_API_KEY"),
+    model="claude-3-5-sonnet-latest",
+    system_prompt="You are a helpful assistant.",
+)
+
 
 # Set some Shiny page options
 ui.page_opts(
@@ -28,18 +37,5 @@ chat.ui()
 # Generate a response when the user submits a message
 @chat.on_user_submit
 async def _(message):
-    response = llm.response_generator(message)
+    response = chat_model.stream(message)
     await chat.append_message_stream(response)
-
-
-# Create an Anthropic Claude chat model
-#
-# Either explicitly set the ANTHROPIC_API_KEY environment variable before launching the
-# app, or set them in a file named `.env`. The `python-dotenv` package will load `.env`
-# as environment variables which can later be read by `os.getenv()`.
-load_dotenv()
-llm = AnthropicChat(
-    api_key=os.environ.get("ANTHROPIC_API_KEY"),
-    model="claude-3-5-sonnet-20240620",
-    system_prompt="You are a helpful assistant.",
-)