+++ shiny/templates/chat/hello-providers/langchain/app.py
@@ -7,10 +7,20 @@
 import os
 
 from app_utils import load_dotenv
-from chatlas import LangChainChat
 from langchain_openai import ChatOpenAI
 
 from shiny.express import ui
+
+# Create a LangChain-based chat model with OpenAI that remembers chat history
+#
+# Either explicitly set the OPENAI_API_KEY environment variable before launching the
+# app, or set them in a file named `.env`. The `python-dotenv` package will load `.env`
+# as environment variables which can later be read by `os.getenv()`.
+load_dotenv()
+chat_model = ChatOpenAI(
+    api_key=os.environ.get("OPENAI_API_KEY"),
+    model="gpt-4o",
+)
 
 # Set some Shiny page options
 ui.page_opts(
@@ -30,20 +40,5 @@ chat.ui()
 # Define a callback to run when the user submits a message
 @chat.on_user_submit
 async def _(message):
-    response = llm.response_generator(message)
+    response = chat_model.stream(message)
     await chat.append_message_stream(response)
-
-
-# Create a LangChain-based chat model with OpenAI that remembers chat history
-#
-# Either explicitly set the OPENAI_API_KEY environment variable before launching the
-# app, or set them in a file named `.env`. The `python-dotenv` package will load `.env`
-# as environment variables which can later be read by `os.getenv()`.
-load_dotenv()
-llm = LangChainChat(
-    ChatOpenAI(
-        api_key=os.environ.get("OPENAI_API_KEY"),
-        model="gpt-4o",
-    ),
-    system_prompt="You are a helpful assistant.",
-)