+++ shiny/templates/chat/hello-providers/ollama/app.py
@@ -1,13 +1,15 @@
 # ------------------------------------------------------------------------------------
 # A basic Shiny Chat example powered by Ollama.
-# To run it, you'll need an Ollama server running locally.
-# To download and run the server, see https://github.com/ollama/ollama
-# To install the Ollama Python client, see https://github.com/ollama/ollama-python
 # ------------------------------------------------------------------------------------
 
-from chatlas import OllamaChat
+from chatlas import ChatOllama
 
 from shiny.express import ui
+
+# ChatOllama() requires an Ollama model server to be running locally.
+# See the docs for more information on how to set up a local Ollama server.
+# https://posit-dev.github.io/chatlas/reference/ChatOllama.html
+chat_model = ChatOllama(model="llama3.1")
 
 # Set some Shiny page options
 ui.page_opts(
@@ -27,9 +29,5 @@ chat.ui()
 # Generate a response when the user submits a message
 @chat.on_user_submit
 async def _(message):
-    response = llm.response_generator(message)
+    response = chat_model.stream(message)
     await chat.append_message_stream(response)
-
-
-# Assumes Ollama is running with llama3.1 available locally
-llm = OllamaChat(model="llama3.1")