+++ shiny/templates/chat/hello-providers/openai/app.py
@@ -1,14 +1,23 @@
 # ------------------------------------------------------------------------------------
-# A basic Shiny Chat example powered by OpenAI's GPT-4o model.
-# To run it, you'll need OpenAI API key.
-# To get setup, follow the instructions at https://platform.openai.com/docs/quickstart
+# A basic Shiny Chat example powered by OpenAI.
 # ------------------------------------------------------------------------------------
 import os
 
 from app_utils import load_dotenv
-from chatlas import OpenAIChat
+from chatlas import ChatOpenAI
 
 from shiny.express import ui
+
+# ChatOpenAI() requires an API key from OpenAI.
+# See the docs for more information on how to obtain one.
+# https://posit-dev.github.io/chatlas/reference/ChatOpenAI.html
+load_dotenv()
+chat_model = ChatOpenAI(
+    api_key=os.environ.get("OPENAI_API_KEY"),
+    model="gpt-4o",
+    system_prompt="You are a helpful assistant.",
+)
+
 
 # Set some Shiny page options
 ui.page_opts(
@@ -28,18 +37,5 @@ chat.ui()
 # Generate a response when the user submits a message
 @chat.on_user_submit
 async def _(message):
-    response = llm.response_generator(message)
+    response = chat_model.stream(message)
     await chat.append_message_stream(response)
-
-
-# Create a chatlas OpenAI chat model
-#
-# Either explicitly set the OPENAI_API_KEY environment variable before launching the
-# app, or set them in a file named `.env`. The `python-dotenv` package will load `.env`
-# as environment variables which can later be read by `os.getenv()`.
-load_dotenv()
-llm = OpenAIChat(
-    api_key=os.environ.get("OPENAI_API_KEY"),
-    model="gpt-4o",
-    system_prompt="You are a helpful assistant.",
-)