+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * A couple new features for `ui.Chat()`:
     * The `on_user_submit` decorator (of a `ui.Chat()` instance) now passes user input to the decorated function. This reduces the need to call `.messages()` to get the message history for response generation. (#1690)
     * `ui.Chat()` now integrates with the `chatlas` package, simplifying the response generation process for packages such as `openai`, `anthropic`, etc. (#1690)
-    * New templates demonstrating tool calling and response generation with `chatlas` are available via `shiny create`. (#1690)
+    * New templates demonstrating response generation with `chatlas` are available via `shiny create`. (#1690)
 
 ### Breaking changes